+++ plugins/rplanners/parabolicsmoother.cpp
@@ -117,7 +117,12 @@ public:
             dynamicpath.SetJointLimits(parameters->_vConfigLowerLimit,parameters->_vConfigUpperLimit);
             dynamicpath.SetMilestones(path);   //now the trajectory starts and stops at every milestone
             RAVELOG_DEBUG(str(boost::format("initial path size %d, duration: %f")%path.size()%dynamicpath.GetTotalTime()));
-            ParabolicRamp::RampFeasibilityChecker checker(this,parameters->_pointtolerance);
+
+            ParabolicRamp::Vector tol = parameters->_vConfigResolution;
+            FOREACH(it,tol) {
+                *it *= parameters->_pointtolerance;
+            }
+            ParabolicRamp::RampFeasibilityChecker checker(this,tol);
             int numshortcuts=0;
             if( !!_parameters->_setstatefn ) {
                 dynamicpath.Shortcut(parameters->_nMaxIterations,checker);
@@ -210,14 +215,37 @@ public:
 
     virtual bool ConfigFeasible(const ParabolicRamp::Vector& a)
     {
-        _parameters->_setstatefn(a);
-        return _parameters->_checkpathconstraintsfn(a,a,IT_OpenStart,PlannerBase::ConfigurationListPtr());
+        // have to also test with tolerances!
+        boost::array<dReal,3> perturbations = {{ 0,_parameters->_pointtolerance,-_parameters->_pointtolerance}};
+        ParabolicRamp::Vector anew(a.size());
+        FOREACH(itperturbation,perturbations) {
+            for(size_t i = 0; i < a.size(); ++i) {
+                anew[i] = a[i] + *itperturbation * _parameters->_vConfigResolution.at(i);
+            }
+            _parameters->_setstatefn(anew);
+            if( !_parameters->_checkpathconstraintsfn(a,a,IT_OpenStart,PlannerBase::ConfigurationListPtr()) ) {
+                return false;
+            }
+        }
+        return true;
     }
 
     virtual bool SegmentFeasible(const ParabolicRamp::Vector& a,const ParabolicRamp::Vector& b)
     {
-        _parameters->_setstatefn(a);
-        return _parameters->_checkpathconstraintsfn(a,b,IT_OpenStart,PlannerBase::ConfigurationListPtr());
+        // have to also test with tolerances!
+        boost::array<dReal,3> perturbations = {{ 0,_parameters->_pointtolerance,-_parameters->_pointtolerance}};
+        ParabolicRamp::Vector anew(a.size()), bnew(b.size());
+        FOREACH(itperturbation,perturbations) {
+            for(size_t i = 0; i < a.size(); ++i) {
+                anew[i] = a[i] + *itperturbation * _parameters->_vConfigResolution.at(i);
+                bnew[i] = b[i] + *itperturbation * _parameters->_vConfigResolution.at(i);
+            }
+            _parameters->_setstatefn(anew);
+            if( !_parameters->_checkpathconstraintsfn(anew,bnew,IT_OpenStart,PlannerBase::ConfigurationListPtr()) ) {
+                return false;
+            }
+        }
+        return true;
     }
 
     virtual dReal Rand()